% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_exe_request}
\alias{db_exe_request}
\title{Execute YAML Change Request}
\usage{
db_exe_request(request.filename)
}
\arguments{
\item{request.filename}{Path to the YAML request file}
}
\description{
Execute YAML Change Request
}
