% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_write}
\alias{db_write}
\title{Write Table}
\usage{
db_write(tbl, table, msg, request.id, dbconn = .dbconn())
}
\arguments{
\item{tbl}{Table object to write.}

\item{table}{Table name.}

\item{msg}{Change message.}

\item{request.id}{Request ID associated to this change.}
}
\description{
Write Table
}
