% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bitdecode}
\alias{.bitdecode}
\title{Decode decimal number into binary.}
\usage{
.bitdecode(decimal, i)
}
\arguments{
\item{decimal}{A vector of decimals.}

\item{i}{Numeric. The position in decoded binary string to extract.}
}
\description{
Decode decimal number into binary.
}
