% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bitencode}
\alias{.bitencode}
\title{Encode binary results into decimal}
\usage{
.bitencode(input)
}
\arguments{
\item{input}{A numeric vector, a logical vector or a logical matrix of binary bits to be encoded. Columns of the logical matrix represent the objects and rows the bits.}
}
\value{
Numeric vector. The converted decimal number.
}
\description{
Encode binary results into decimal
}
