% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache_clean}
\alias{.cache_clean}
\title{Clean Stale Cache Files}
\usage{
.cache_clean(period = 7, pkg = utils::packageName())
}
\arguments{
\item{period}{Numeric value indicating the maximum age of cache files in days.
Files older than this will be deleted. Defaults to \code{7}.}

\item{pkg}{Package name for the cache directory. Defaults to the current package environment.}
}
\value{
Invisibly returns the number of files deleted.
}
\description{
Deletes cache files that have not been modified within the specified period.
}
