% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache_read}
\alias{.cache_read}
\title{Read Object from Cache}
\usage{
.cache_read(name, identifier)
}
\arguments{
\item{name}{Character string representing the conventional name of the object.}

\item{identifier}{A vector or list used to generate the unique hash.}
}
\value{
The cached R object.
}
\description{
Reads a previously cached \code{.rds} file based on its name and identifier hash.
}
\details{
Throws an error if the corresponding cache file does not exist.
}
