% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache_write}
\alias{.cache_write}
\title{Write Object to Cache}
\usage{
.cache_write(obj, name, identifier)
}
\arguments{
\item{obj}{The R object to cache.}

\item{name}{Character string representing the conventional name of the object.}

\item{identifier}{A vector or list used to generate a unique hash, distinguishing
different versions of the object under the same \code{name}.}
}
\value{
The original object \code{obj}, invisibly.
}
\description{
Serializes an R object to an OS-specific cache directory as an \code{.rds} file.
}
