% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{.dbconn}
\alias{.dbconn}
\title{Connect to SQLite database}
\usage{
.dbconn(path = NULL)
}
\arguments{
\item{path}{Path to SQLite database. If not given, configuration file will be searched.}
}
\description{
Connect to SQLite database
}
