% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dependency_check}
\alias{.dependency_check}
\title{Check whether the required dependency is installed}
\usage{
.dependency_check(pkg, install.call = NULL)
}
\arguments{
\item{pkg}{Package name.}

\item{install.call}{A suggested call for installing the package if missing. The call must be passed quoted by \code{\link[=quote]{quote()}} or \code{\link[rlang:expr]{rlang::expr()}}.}
}
\description{
Check whether the required dependency is installed
}
