% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dir_check}
\alias{.dir_check}
\title{Check output directory}
\usage{
.dir_check(dir, overwrite = FALSE, interactive = TRUE)
}
\arguments{
\item{dir}{A path to output directory.}

\item{overwrite}{Overwrite permission. Default to \code{FALSE}}

\item{interactive}{Interactive console request on user for granting overwrite permission.}
}
\description{
Check output directory
}
