% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.get_cache_path}
\alias{.get_cache_path}
\title{Get Cache File Path}
\usage{
.get_cache_path(name, identifier, pkg = utils::packageName())
}
\arguments{
\item{name}{Character string for the object's base name.}

\item{identifier}{An arbitrary vector or list to hash.}

\item{pkg}{Package name for the cache directory. Defaults to the current package environment.}
}
\value{
Character string of the absolute file path.
}
\description{
Internal helper to generate a deterministic file path for cached objects.
}
\keyword{internal}
