% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file.R
\name{.is_file}
\alias{.is_file}
\title{Check if paths exists and are files.}
\usage{
.is_file(path, error = F)
}
\arguments{
\item{path}{Character vector. Path(s) to be tested.}

\item{error}{Logical. Whether throw an error if any path didn't pass the text. Default to \code{FALSE}.}
}
\value{
A logical vector in same length of \code{path}.
}
\description{
Check if paths exists and are files.
}
