% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.overwrite_check}
\alias{.overwrite_check}
\title{Check if target folder exists and has files}
\usage{
.overwrite_check(dir, overwrite, interactive)
}
\value{
Numeric code for overwrite permission. 0: user-defined overwrite or target folder does not yet exist. 1: user-specified overwrite rejection. 2: overwrite rejected and no options given for permission request.
}
\description{
Check if target folder exists and has files
}
