% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safe_name}
\alias{.safe_name}
\title{Turn Characters Vectors to be Safe}
\usage{
.safe_name(c, replacement = "_", allow = "")
}
\arguments{
\item{c}{Character vector. Input to be convert into safe characters.}

\item{replacement}{Character vector of length 1. Element to replace unsafe characters}

\item{allow}{Regex expression. To be added as \verb{[^allow]}. Matching additional characters to be retained. Default to none, which only allows safest characters (alphabet, numbers, underscore).}
}
\value{
Character vector of length \code{c}. Safe version of the input. Attribute \code{lut} provides a look-up table of the conversion, which can be used to map the original value to the safe values.
}
\description{
Turn Characters Vectors to be Safe
}
