% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safe_write}
\alias{.safe_write}
\title{Safely write output files with overwrite check}
\usage{
.safe_write(
  obj,
  ...,
  dir = NULL,
  file = NULL,
  files = NULL,
  write_func = NULL,
  create_dir = TRUE,
  write_func_args = NULL,
  overwrite_args = list(interactive = interactive()),
  csv_args = list(sep = ";", row.names = FALSE)
)
}
\arguments{
\item{obj}{Object to write.}

\item{...}{Placeholder. Nothing will be evaluated in this slot. It enforces whoever use this function to contiously specifying their output as dir, file, or files.}

\item{dir}{Directory to write the output files}

\item{file}{Mantadory if \code{dir} is provided. File name of the output.}

\item{files}{File path to write the object.}

\item{write_func}{Function to write object, the first argument must be object to write and the second argument must be the path. Default to \code{NULL}.}

\item{create_dir}{Logical. Whether to create directories if the target directories do not exist. If set to \code{TRUE}, the directories will be created recursively until the file can be written. Default to \code{TRUE}.}

\item{write_func_args}{Arguments to passed to \code{write_func}.}

\item{overwrite_args}{Arguments passed to \code{\link[=.overwrite_check]{.overwrite_check()}}}

\item{csv_args}{Arguments passed to \code{\link[=write.table]{write.table()}}}
}
\description{
Write object to output paths. There are three ways to specify output paths. Provide base file names for all objects and a output directory for all, provide base file names and output directories for all objects, or provide file paths for all objects. The resulted file paths must be in the same length of the object.
}
\details{
Write function could be specified in \code{write_func}. If set to \code{NULL}, some file extension will be detected and written with corresponding functions.
}
