% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.say_no}
\alias{.say_no}
\title{Gracefully say no}
\usage{
.say_no(verbose = cat)
}
\arguments{
\item{verbose}{A function to say no. Default to \code{cat}. Set to \code{NULL} to silently return no without saying it out loud.}
}
\value{
A variety of ways to say no.
}
\description{
Powered by hotheadhacker/no-as-a-service (https://github.com/hotheadhacker/no-as-a-service).
}
\seealso{
\code{\link[=.get_no]{.get_no()}}
}
