% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.use_cache}
\alias{.use_cache}
\title{Evaluate Expression with Caching}
\usage{
.use_cache(name, identifier, expr)
}
\arguments{
\item{name}{Character string representing the conventional name of the object.}

\item{identifier}{A vector or list used to generate the unique hash.}

\item{expr}{An expression to evaluate if the cache is missing.}
}
\value{
The result of \code{expr}, either retrieved from cache or newly evaluated.
}
\description{
Attempts to retrieve an object from the cache. If it is not found,
it evaluates the provided expression, writes the result to the cache, and returns it.
}
