% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.user_input_select}
\alias{.user_input_select}
\title{Ask user for input by selecting given options.}
\usage{
.user_input_select(prompt, options, random = T, abort_option = T)
}
\arguments{
\item{options}{Character vectors. Options.}

\item{random}{Logical. Randomize option order. Default to \code{TRUE}}
}
\description{
Ask user for input by selecting given options.
}
