% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.with_timeout}
\alias{.with_timeout}
\title{Abort function after time out}
\usage{
.with_timeout(expr, cpu, elapsed)
}
\description{
Abort function after time out
}
\details{
Source: https://stackoverflow.com/a/53018594
}
\author{
landau
}
