% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_best_match}
\alias{find_best_match}
\title{Find Best Match in A Vector of Strings using Automatic Gap Detection}
\usage{
find_best_match(pattern, strings, index = F, silent = F)
}
\arguments{
\item{pattern}{A pattern to look for.}

\item{strings}{A vector of strings to look for the pattern.}

\item{index}{Logical. Whether to return index of the best matching item in the string vector (TRUE) or return the best matching item value (FALSE).}

\item{silent}{Logical. Set to TRUE to disable result printing.}
}
\description{
Find Best Match in A Vector of Strings using Automatic Gap Detection
}
\author{
Zhehao Hu
}
