% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{paletten}
\alias{paletten}
\title{Generate a color palette from base palette}
\usage{
paletten(palette, n)
}
\arguments{
\item{palette}{Base palette in hex code to interpolate.}

\item{n}{Number of colors to generate.}
}
\value{
A vector of colors.
}
\description{
Generate a color palette from base palette
}
