% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{pub_nodescription}
\alias{pub_nodescription}
\title{Remove Descriptions on \code{ggplot} Plots for Publication}
\usage{
pub_nodescription(ggobj)
}
\description{
Remove Descriptions on \code{ggplot} Plots for Publication
}
