% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file.R
\name{remove_empty_folders}
\alias{remove_empty_folders}
\title{Remove Empty Folders}
\usage{
remove_empty_folders(path)
}
\arguments{
\item{path}{Path to directory.}
}
\description{
Removes empty folders recursively downwards.
}
