% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{themes}
\alias{themes}
\alias{theme_monochrome}
\alias{theme_monochrome_open}
\alias{theme_pub}
\title{HPLNC Graphic Themes}
\usage{
theme_monochrome(...)

theme_monochrome_open(...)

theme_pub(...)
}
\arguments{
\item{...}{Further argument to be passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
HPLNC Graphic Themes
}
