% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file.R
\name{update_md_frontmatter}
\alias{update_md_frontmatter}
\title{Update a markdown file with new YAML frontmatter}
\usage{
update_md_frontmatter(
  fm,
  file_path,
  exclude_property = NULL,
  mode = c("modify", "rewrite")
)
}
\arguments{
\item{fm}{The modified frontmatter list}

\item{file_path}{Path to the markdown file}

\item{exclude_property}{Character vector of keys to strip before writing}

\item{mode}{Update mode. Unambiguous abbreviations are accepted. Default to \code{modify}, which keeps the unmodified properties. Also accept \code{rewrite}, which ignores original properties and rewrite completely with given \code{fm}.}
}
\description{
Update a markdown file with new YAML frontmatter
}
