% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhutilsr-package.R
\docType{package}
\name{zhutilsr-package}
\alias{zhutilsr}
\alias{zhutilsr-package}
\title{zhutilsr: Zhehao's Utilities}
\description{
Zhehao's Utilities
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zzzhehao/zhutilsr}
  \item Report bugs at \url{https://github.com/zzzhehao/zhutilsr/issues}
}

}
\author{
\strong{Maintainer}: Zhehao Hu \email{hu_zhehao@hotmail.com} (\href{https://orcid.org/0009-0006-2408-1513}{ORCID})

}
\keyword{internal}
