% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.clean_yaml_list}
\alias{.clean_yaml_list}
\title{Safely remove flat or nested properties from a list}
\usage{
.clean_yaml_list(lst, excludes = NULL)
}
\arguments{
\item{lst}{The frontmatter list}

\item{excludes}{A character vector of properties to remove (e.g., c("name", "parent$child"))}
}
\description{
Safely remove flat or nested properties from a list
}
