% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dependency_check}
\alias{.dependency_check}
\title{Check whether the required dependency is installed}
\usage{
.dependency_check(pkg, install.call = NULL)
}
\arguments{
\item{pkg}{Character vector. Package names.}

\item{install.call}{A suggested function (or list of functions) for installing the package if missing. Default to \code{NULL}, no installation will be suggested.}
}
\value{
A logical vector of the same length as \code{pkg} indicating if packages are installed.
}
\description{
Check whether the required dependency is installed
}
