% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_taxonomy.R
\name{format_organism_name}
\alias{format_organism_name}
\title{format organism label}
\usage{
format_organism_name(
  label,
  tag_open = "<i>",
  tag_close = "</i>",
  plotmath.safe = F
)
}
\description{
format organism label
}
