% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file.R
\name{svg2png}
\alias{svg2png}
\title{Convert a single SVG file to PNG(s) with varying size, color, and opacity}
\usage{
svg2png(input, output_dir, size = 256, alpha = 1, hex = "#000000")
}
\arguments{
\item{input}{Path to the input SVG file.}

\item{output_dir}{Path to the output directory.}

\item{size}{Numeric vector of target sizes in pixels.}

\item{alpha}{Numeric vector of opacity values between 0 and 1.}

\item{hex}{Character/Numeric vector of colors (e.g., "#000000", "red").}
}
\description{
Convert a single SVG file to PNG(s) with varying size, color, and opacity
}
