% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{svg2pngf}
\alias{svg2pngf}
\title{Convert a folder of SVG files to PNGs with varying size, color, and opacity}
\usage{
svg2pngf(
  inputfolder,
  outputfolder,
  size = 256,
  alpha = 1,
  hex = "#000000",
  recursive = FALSE,
  regex = NULL
)
}
\arguments{
\item{inputfolder}{Path to the input directory containing SVGs.}

\item{outputfolder}{Path to the output directory.}

\item{size}{Numeric vector of target sizes in pixels.}

\item{alpha}{Numeric vector of opacity values between 0 and 1.}

\item{hex}{Character/Numeric vector of colors (e.g., "#000000", "red").}

\item{recursive}{Logical. Should sub-directories be processed while retaining folder structure?}

\item{regex}{Character. Optional regex string to filter the target SVG files.}
}
\description{
Convert a folder of SVG files to PNGs with varying size, color, and opacity
}
